FortuneGraphに内容を移動
